refactor: useSearch hook
