+++ src/screens/GlobalSearchScreen/GlobalSearchScreen.tsx
@@ -5,7 +5,7 @@ import { ProgressBar } from 'react-native-paper';
 import { EmptyView, SearchbarV2 } from '@components/index';
 import GlobalSearchResultsList from './components/GlobalSearchResultsList';
 
-import useSearch from '@hooks/useSearch';
+import { useSearch } from '@hooks';
 import { useTheme } from '@hooks/useTheme';
 
 import { getString } from '@strings/translations';
